Move and refactor checkTrades to BotUtils class
The checkTrades function was moved from OldCommands/Utils.php (now deleted) to the BotUtils class, refactored as a method. Main.php was updated to call the new method, and minor improvements were made such as updating file paths and using class properties.
