+++ Main.php
@@ -45,6 +45,7 @@
 		
 		$discord->getLoop()->addPeriodicTimer(120, function () use ($utils) {
 			$utils->checkDota();
+			$utils->checkTrades();
 		});
 		
 		$discord->getLoop()->addPeriodicTimer(300, function () use ($utils) {
